requirements.txt

python-telegram-bot==20.3
gspread==5.11.1
google-auth==2.25.0
